requirements.txt
Django==5.0.7
djangorestframework==3.15.2

djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
cryptography==41.0.7

psycopg2-binary==2.9.9
django-extensions==3.2.3

drf-spectacular==0.27.2

Pillow==10.0.1
django-storages==1.14.2

python-decouple==3.8
django-environ==0.11.2

marshmallow==3.20.1
django-filter==23.3

redis==5.0.1
django-redis==5.4.0
celery==5.3.4

sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

flake8==6.1.0
black==23.11.0
isort==5.12.0

ipython==8.17.2
django-seed==0.3.1

scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2

pinecone-client==2.2.4
openai==1.3.7

django-anymail==10.2
python-dotenv